+++ petshop-app/src/MyApp-PetShop/api.js
@@ -1,0 +1,310 @@
+export const listDog = [
+  {
+    typeOf: "Alaska",
+    name: "ALASKA NGÁO",
+    priceOld: 10500000,
+    priceCurrent: 9000000,
+    url: "http://thucanhviet.com/wp-content/uploads/2017/09/cho-alaska-mau-nau-do.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA NGẦU",
+    priceCurrent: 12500000,
+    url: "https://sieupet.com/sites/default/files/alaka2_0.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA CHẤT",
+    priceCurrent: 11500000,
+    url: "http://www.thukieng.com/wp-content/uploads/2016/08/ban-cho-alaska-nau-do-17-3.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA XINH",
+    priceOld: 11000000,
+    priceCurrent: 9500000,
+    url: "https://i.pinimg.com/originals/a4/d0/df/a4d0df010b577941a884e3861e139f4f.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA SIÊU CUTE",
+    priceCurrent: 9500000,
+    url: "https://tapchithucung.vn/Uploads/images/content/nhung-su-that-thu-vi-ve-cho-alaska-it-nguoi-biet.JPG",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA ĐÁNG YÊU",
+    priceCurrent: 13500000,
+    url: "https://toplist.vn/images/800px/traicholetrungcom-479111.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA ĐEN TRẮNG",
+    priceOld: 14500000,
+    priceCurrent: 1000000,
+    url: "https://matpetfamily.com/wp-content/uploads/2020/03/3.3.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA NÂU ĐỎ",
+    priceCurrent: 17500000,
+    url: "https://dogily.vn/wp-content/uploads/2021/12/alaska-nau-do-10-600x400.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA XÁM TRẮNG",
+    priceOld: 16000000,
+    priceCurrent: 1100000,
+    url: "https://i.pinimg.com/736x/29/dd/31/29dd31fed0a07c9d4d1ee12085d34425.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA HỒNG PHẤN",
+    priceCurrent: 19000000,
+    url: "https://matpetfamily.com/wp-content/uploads/2020/02/85123850_1551825188298893_9190359651202367488_o.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA Ú XÁM TRẮNG",
+    priceCurrent: 20000000,
+    url: "https://azpet.b-cdn.net/wp-content/uploads/2021/08/Alaska-Oversize-Xam-Trang-C12416-1.jpg",
+  },
+  {
+    typeOf: "Alaska",
+    name: "ALASKA Ú NU",
+    priceCurrent: 22000000,
+    url: "https://matpetfamily.com/wp-content/uploads/2020/08/117886387_1746295572185186_8084319773672697470_o.jpg",
+  },
+
+  {
+    typeOf: "Poodle",
+    name: "POODLE LAI NÂU",
+    priceCurrent: 8000000,
+    url: "https://static2.enbaccdn.com/thumb_wm/400/ebx/2020/07/19/1106745/15951408059056303326.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE LAI TRẮNG",
+    priceOld: 7000000,
+    priceCurrent: 600000,
+    url: "https://thucanh.net/wp-content/uploads/2021/07/sua-tam-cho-cho-poodle-trang.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE DỄ THƯƠNG",
+    priceCurrent: 8500000,
+    url: "https://thucanh.net/wp-content/uploads/2021/07/sua-tam-cho-cho-poodle-trang.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE XINH",
+    priceOld: 9400000,
+    priceCurrent: 8400000,
+    url: "https://petviet.vn/wp-content/uploads/2018/04/hinh-anh-cho-toy-poodle-07.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE SIÊU CUTE",
+    priceCurrent: 9500000,
+    url: "https://poodlefamily.com/wp-content/uploads/2021/10/POODLE-TRANG-01.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE ĐÁNG YÊU",
+    priceCurrent: 8200000,
+    url: "https://traichomeo.com/wp-content/uploads/2022/03/cho-tyty-3.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE NÂU",
+    priceOld: 10300000,
+    priceCurrent: 900000,
+    url: "http://poodledep.com/poodle/poodle-nau-2-thang.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE NÂU TRẮNG",
+    priceCurrent: 15500000,
+    url: "https://cdn.chotot.com/laMPrqIYIkIhJSpNZKxL246DYRsxEH5RwLJ7naZXdYA/preset:listing/plain/8a47628130181660bf7ea58c27189ca2-2789090655283620449.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE SIZE LỚN",
+    priceOld: 12300000,
+    priceCurrent: 10500000,
+    url: "https://cdn.chotot.com/L6Ac6S1tsC-gpXY4n68uRcwJhUKgrRsbfJ9BxTl5fa0/preset:listing/plain/474c4ea1be1e814cf2a9b780152d1921-2788693868720216353.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE ĐEN CHẤT",
+    priceCurrent: 13400000,
+    url: "https://dogily.vn/wp-content/uploads/2022/06/poodle-mau-den-duc-5-400x400.png",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE COUPLE SIÊU NGẦU",
+    priceCurrent: 35900000,
+    url: "https://thucanh.net/wp-content/uploads/2021/05/Cho-poodle-28.jpg",
+  },
+  {
+    typeOf: "Poodle",
+    name: "POODLE SIZE MINI",
+    priceCurrent: 18000000,
+    url: "https://cdn.chotot.com/zeVjkkdQtJRRd4kBz5WZgb0JeaMHuWsFPkm0B7dXRhE/preset:listing/plain/0adc9fddec2e5c1709ca0975c5eec4b7-2776527987955621828.jpg",
+  },
+];
+
+export const listCat = [
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR CHẤT",
+    priceCurrent: 11000000,
+    url: "https://petonesaigon.com/wp-content/uploads/2021/08/10-tieu-chuan-can-biet-khi-mua-meo-anh-long-dai-british-longhair.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR COUPLE TRẮNG",
+    priceCurrent: 40000000,
+    url: "https://www.petcim.com/pictures_ilan/mega_dfeae745f3e10987bbbbe5669f0f7b0a.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR TRẮNG XÁM",
+    priceCurrent: 13000000,
+    url: "https://upload.wikimedia.org/wikipedia/commons/thumb/2/2a/British_Longhair_-_Blue_Bicolor.jpg/800px-British_Longhair_-_Blue_Bicolor.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR MẮT LẠ",
+    priceCurrent: 19000000,
+    url: "https://dogily.vn/wp-content/uploads/2022/03/meo-anh-long-dai-ald-nhap-dogily-2.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR PHỐI MÀU",
+    priceCurrent: 10000000,
+    url: "https://richardshomecattery.com/wp-content/uploads/2022/06/1-11.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR MẮT 2 MÀU",
+    priceOld: 20000000,
+    priceCurrent: 19000000,
+    url: "https://dogily.vn/wp-content/uploads/2022/06/ald-hai-mau-mat-duc-2.png",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR COUPLE XÁM TRẮNG",
+    priceCurrent: 30000000,
+    url: "https://blogchomeo.com/meo-anh-long-dai/british-longhair-tabby/",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR SIÊU ĐẸP",
+    priceOld: 17000000,
+    priceCurrent: 16000000,
+    url: "https://i.pinimg.com/736x/93/22/56/932256b3ca3f8966526683cf01cbc964.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR DỄ THƯƠNG",
+    priceCurrent: 16000000,
+    url: "https://www.thukieng.com/wp-content/uploads/2019/01/meo-anh-mau-sieu-dep-nua-cuoi-nam-2021-thu-kieng-4-300x280.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR ĐÁNG YÊU",
+    priceOld: 16500000,
+    priceCurrent: 14500000,
+    url: "https://www.thukieng.com/wp-content/uploads/2021/06/alnt9-300x300.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR SIÊU XINH",
+    priceCurrent: 16500000,
+    url: "https://fusiongroup.vn/upload/file/thumb_800x0/meo-anh-long-dai-noi-ban-va-gia-meo-ald-fusion-1646297043.jpg",
+  },
+  {
+    typeOf: "British Longhair",
+    name: "BRITISH LONGHAIR SIÊU XỊN",
+    priceOld: 17000000,
+    priceCurrent: 15000000,
+    url: "https://file.hstatic.net/200000108863/file/1_78c5c64d1ac14c80922efd9f0df2e830_grande.png",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN ĐÁNG YÊU",
+    priceCurrent: 9000000,
+    url: "https://azpet.b-cdn.net/wp-content/uploads/2021/07/Munchkin-Trang-M12208-1-1024x1024.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN DỄ THƯƠNG",
+    priceOld: 10000000,
+    priceCurrent: 9000000,
+    url: "https://azpet.b-cdn.net/wp-content/uploads/2021/07/Munchkin-Trang-M12208-2-375x375.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN SIÊU CHẤT",
+    priceCurrent: 17000000,
+    url: "https://dogily.vn/wp-content/uploads/2021/09/meo-munchkin-scottish-chan-ngan-tai-cup-7-400x400.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN NGẦU",
+    priceOld: 18000000,
+    priceCurrent: 17000000,
+    url: "https://dogily.vn/wp-content/uploads/2021/09/meo-chan-ngan-tai-cup-xam-xanh.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN VÀNG TRẮNG",
+    priceCurrent: 1100000,
+    url: "https://static.chotot.com/storage/chotot-kinhnghiem/c2c/2019/10/meo-munchkin-21-1024x856.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN ĐEN",
+    priceCurrent: 12000000,
+    url: "https://azpet.b-cdn.net/wp-content/uploads/2021/07/Munchkin-Xanh-Xam-M12140-1-300x300.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN NÂU",
+    priceOld: 13000000,
+    priceCurrent: 11000000,
+    url: "https://thegioiloaimeo.com/wp-content/uploads/2020/02/meo-munchkin2.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN XỊN",
+    priceCurrent: 15000000,
+    url: "https://petmaster.vn/petroom/wp-content/uploads/2020/04/meo-munchkin-chan-ngan.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN SIÊU XINH",
+    priceOld: 12000000,
+    priceCurrent: 10000000,
+    url: "https://azpet.b-cdn.net/wp-content/uploads/2021/07/Munchkin-Socola-M12143-1-1024x1024.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN Ú",
+    priceCurrent: 19000000,
+    url: "https://afamilycdn.com/150157425591193600/2020/6/2/photo-1-15911072478101340790046-1591107552666-1591107553285684659439.jpg",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN MẶT NGẦU",
+    priceCurrent: 18000000,
+    url: "https://lh3.googleusercontent.com/PNerwLuncWANRYRTeiZWCF9gruWh_nFpsu_hGGIG-Gvzn1SuNCEVZHgukztcdgqhIPGYrFr6O3Y3iYjwJ50RH9lxVhoi6xW93ZdlQqvGkI-5bnyp6BkW-cAxFfaZA6NKMA5gaA",
+  },
+  {
+    typeOf: "Munchkin",
+    name: "MUNCHKIN SIÊU CHẤT",
+    priceCurrent: 20000000,
+    url: "https://azpet.b-cdn.net/wp-content/uploads/2021/07/Munchkin-Bicolor-M12205-1-1024x1024.jpg",
+  },
+];
