+++ petshop-app/src/MyApp-PetShop/component/Banner.jsx
