+++ petshop-app/src/MyApp-PetShop/css/listpet.css
@@ -22,7 +22,6 @@
   width: 90%;
   height: fit-content;
 }
-
 .card_pet:hover {
   transform: scale(1.1);
 }
