+++ petshop-app/src/MyApp-PetShop/css/searchDog.css
@@ -1,0 +1,107 @@
+
+.wrapper {
+    width: 96%;
+    max-width: 1140px;
+    margin: 0 auto;
+}
+/* search input */
+.search-wrapper {
+    margin: 48px 0;
+    display: flex;
+    justify-content: space-between;
+}
+
+@media (max-width: 375px) {
+    .search-input {
+        width: 100%;
+    }
+    .search-wrapper {
+        justify-content: start;
+        flex-wrap: wrap;
+    }
+    .select {
+        margin-top: 3.5em;
+    }
+}
+
+.search-input {
+    background-image: url("data:image/svg+xml;base64,PHN2ZyB3aWR0aD0iMjQiIGhlaWdodD0iMjQiIHhtbG5zPSJodHRwOi8vd3d3LnczLm9yZy8yMDAwL3N2ZyIgZmlsbC1ydWxlPSJldmVub2RkIiBjbGlwLXJ1bGU9ImV2ZW5vZGQiPjxwYXRoIGQ9Ik0xNS44NTMgMTYuNTZjLTEuNjgzIDEuNTE3LTMuOTExIDIuNDQtNi4zNTMgMi40NC01LjI0MyAwLTkuNS00LjI1Ny05LjUtOS41czQuMjU3LTkuNSA5LjUtOS41IDkuNSA0LjI1NyA5LjUgOS41YzAgMi40NDItLjkyMyA0LjY3LTIuNDQgNi4zNTNsNy40NCA3LjQ0LS43MDcuNzA3LTcuNDQtNy40NHptLTYuMzUzLTE1LjU2YzQuNjkxIDAgOC41IDMuODA5IDguNSA4LjVzLTMuODA5IDguNS04LjUgOC41LTguNS0zLjgwOS04LjUtOC41IDMuODA5LTguNSA4LjUtOC41eiIvPjwvc3ZnPg==");
+    background-color: var(--bg-offset);
+    background-size: 16px 16px;
+    background-position: left 10px center;
+    background-repeat: no-repeat;
+    padding: 1.4em 2em;
+    padding-left: 2.7em;
+    border: 1px solid var(--border);
+    color: var(--gray);
+    box-shadow: 0px 4px 6px var(--border);
+    transition: all 0.4s cubic-bezier(0.215, 0.61, 0.355, 1);
+
+
+}
+.search-input:hover {
+    box-shadow: 0px 0px 0px var(--border);
+}
+/* select from moderncss.dev */
+select {
+    
+    appearance: none !important;
+   
+    outline: none;
+    background-color: var(--bg-offset);
+    border-radius: 0.25em;
+    border-width: 1px;
+    border-style: solid;
+    border-color: var(--border);
+    padding: 1.4em 2em 1.4em 1em;
+    margin: 0;
+    margin-right: 1em;
+    width: 100%;
+    font-family: inherit;
+    font-size: inherit;
+    cursor: inherit;
+    line-height: inherit;
+    width: 100%;
+    color: var(--gray);
+}
+.select {
+    min-width: 15ch;
+    max-width: 30ch;
+    cursor: pointer;
+    line-height: 1.1;
+    background-color: transparent;
+    display: grid;
+    grid-template-areas: "select";
+    align-items: center;
+    position: relative;
+    box-shadow: 0px 4px 6px var(--border);
+    transition: all 0.4s cubic-bezier(0.215, 0.61, 0.355, 1);}
+
+    .select:hover {
+        box-shadow: 0px 0px 0px var(--border);
+    }
+
+    .select::after {
+        content: "";
+        display: block;
+        width: 0.8em;
+        height: 0.5em;
+        background-color: var(--text);
+        clip-path: polygon(100% 0%, 0 0%, 50% 100%);
+        justify-self: end;
+        margin-right: 1em;
+    }
+
+select,
+.select:after {
+    grid-area: select;
+}
+select:focus + .focus {
+    position: absolute;
+    top: -1px;
+    left: -1px;
+    right: -1px;
+    bottom: -1px;
+    border: 2px solid var(--text);
+    border-radius: inherit;
+}
