+++ petshop-app/src/MyApp-PetShop/helper/constants.js
@@ -1,0 +1,5 @@
+export const PAGINATION_DOG = {
+    LIMIT_DOG: 4,
+    CURRENT_PAGE_DOG: 1,
+    TOTAL_PAGE_DOG: 0,
+};
