+++ petshop-app/src/MyApp-PetShop/page/Card.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Card() {
+  return (
+    <div>Card</div>
+  )
+}
