+++ petshop-app/src/MyApp-PetShop/page/Checkout.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Checkout() {
+  return (
+    <div>Checkout</div>
+  )
+}
