+++ petshop-app/src/MyApp-PetShop/page/Home.jsx
@@ -1,13 +1,13 @@
 import React from "react";
+import { listCat, listDog } from "../api";
 import Banner from "../component/Banner";
 import PetSaleUp from "../component/PetSaleUp";
 
-export default function Home({allPet}) {
-  console.log("list", allPet)
+export default function Home() {
   return (
     <div>
       <Banner />
-      <PetSaleUp  allPet={allPet}/> 
+      <PetSaleUp listDog={listDog} listCat={listCat} />
     </div>
   );
 }
