+++ petshop-app/src/constants.js
@@ -1,10 +1,5 @@
-export const KEY_SEARCH = {
-    USER_NAME: "name",
-    PRICE_CURRENT: "priceCurrent",
-};
-
 export const PAGINATION = {
-    LIMIT: 2,
+    LIMIT: 4,
     CURRENT_PAGE: 1,
     TOTAL_PAGE: 0,
 };
