--- petshop-app/src/page/Cart.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Cart() {
-  return (
-    <div>Cart</div>
-  )
-}
