--- petshop-app/src/page/Cats.jsx
@@ -1,10 +1,0 @@
-import React from 'react'
-import ListCat from '../component/ListCat'
-
-export default function Cats({allPet}) {
-  return (
-    <div>
-      <ListCat pets={allPet}/>
-    </div>
-  )
-}
