--- petshop-app/src/page/Order.jsx
@@ -1,7 +1,0 @@
-import React from 'react'
-
-export default function Order() {
-  return (
-    <div>Order</div>
-  )
-}
